+++ ui/src/pages/products/ProductModelEdit.tsx
@@ -1233,6 +1233,7 @@ function ProductPanel({
             {/* A single concept exists, you do not have an option to make a new concept */}
             {product.concept && (
               <ExistingConceptDropdown
+                productModel={productModel}
                 product={product}
                 branch={branch}
                 control={control}